requirements.txt
# MinimalERP - Core Requirements (Minimal)

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# HTTP & Requests
httpx==0.26.0
requests==2.31.0

# AI/ML (Minimal - for services)
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
prophet==1.1.5

# OCR & Document Processing
pytesseract==0.3.10
Pillow==10.2.0

# Validation & Parsing
python-dateutil==2.8.2
pytz==2024.1

# Development
python-dotenv==1.0.1

# Monitoring
prometheus-client==0.19.0

# Redis (for future use)
redis==5.0.1
